+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 (v104) - 2026-08-07
+
+## Changed
+
+- Refactor extension to use the new Formatter API
+
+## Fixed
+
+- Fixed `Format with Prettier` command not working when turned off in formatter settings
+
 ## 1.0.0 (v101) - 2026-07-12
 
 ### Added